Fix gacha black boxing on owl pellets
These changes help the bot deal with gachas that are black boxed on owl pellets. The bot will take the owl pellets from the gacha before attempting to transfer Plant Y traps. Then it transfers the pellets back to the gacha. When a gacha is black boxed the drop all bag will now be excess owl pellets instead of plant y.