Added double check for being able to deposit
When collecting crystals it will now walk forward until it sees it can deposit into a dedi. Then it waits 2 seconds and checks it can still deposit. This is to prevent the bot opening crystals too far back from the dedis and getting stuck. This is most likely when a world save hits mid-crystal collection.