Adjusted crystal collection
The bot is checking if it can access the dedis before depositing. In situations where the bot lags back it will walk forward again until it can access the bottom dedis. It needed to also take a step forward once it can access the dedis or sometimes it would fail to reach the top dedis.